Conditional `Restorable#respond_to_missing` impl.
It seems this method is only really needed on Ruby 2.3. And in Ruby 2.5 it
causes a bug with `#flatten` implicitly calling the private `#to_ary` method, as
exemplified here:

    ```
    class A
      def respond_to?(method, include_all = false)
        ::Kernel.puts "respond_to?(#{method.inspect}, #{include_all.inspect})"
      end
    end

    class B
      def respond_to_missing?(method, include_all = false)
        ::Kernel.puts "respond_to_missing?(#{method.inspect}, " \
          "#{include_all.inspect})"
      end
    end

    puts RUBY_VERSION
    [A.new, B.new].flatten

    # Output:
    #
    # 2.4.2
    # respond_to?(:to_ary, true)
    # respond_to_missing?(:to_ary, false)
    #
    # 2.5.0
    # respond_to?(:to_ary, true)
    # respond_to_missing?(:to_ary, true)
    ```

I haven't dug deep enough into this to say I know what's really going on. And I
am not sure I really want to.